refactor Component Styles into Sass
