+++ components/Header.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import Link from "next/link";
 import { useRouter } from "next/router";
-import Styles from "../styles/Components.module.scss"
+import Styles from "../styles/Components/Header.module.scss"
 
 const Header: React.FC = () => {
   const router = useRouter();
@@ -9,52 +9,21 @@ const Header: React.FC = () => {
     router.pathname === pathname;
 
   let left = (
-    <div className="left">
+    <div className={Styles.left}>
       <Link href="/">
-        <a className="bold" data-active={isActive("/")}>
-          <img src="https://i.imgur.com/NxwZ9x0.png"/>
+        <a className={Styles.boldA} data-active={isActive("/")}>
+          <img src="https://i.imgur.com/NxwZ9x0.png" width="100px" height="100px"/>
         </a>
       </Link>
-      <style jsx>{`
-        .bold {
-          font-weight: bold;
-        }
-
-        a {
-          text-decoration: none;
-          color: #000;
-          display: inline-block;
-        }
-
-        img {
-          width: 100px;
-          height: 100px;
-        }
-
-        .left a[data-active="true"] {
-          color: gray;
-        }
-
-        a + a {
-          margin-left: 1rem;
-        }
-      `}</style>
     </div>
   );
 
   let right = null;
 
   return (
-    <nav>
+    <nav className={Styles.nav}>
       {left}
       {right}
-      <style jsx>{`
-        nav {
-          display: flex;
-          padding: 2rem;
-          align-items: center;
-        }
-      `}</style>
     </nav>
   );
 };
