+++ styles/Components/Layout.module.scss
@@ -1,0 +1,3 @@
+.layout {
+  padding: 0 2rem;
+}
