+++ styles/Components/Post.module.scss
@@ -1,0 +1,11 @@
+.post {
+  color: inherit;
+  padding: 2rem;
+}
+.posth2 {
+  text-align: center;
+}
+.authorName {
+  text-align: center;
+  font-size: 13.333px;
+}
