+++ styles/Index.module.scss
@@ -1,16 +1,6 @@
 /* SCSS */
 @import 'colors';
 
-// body {
-//   margin: 0;
-//   padding: 0;
-//   font-size: 16px;
-//   font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-//     Helvetica, Arial, sans-serif, "Apple Color Emoji", "Segoe UI Emoji",
-//     "Segoe UI Symbol";
-//   background: #c5e2de;
-// }
-
 .pageh1 {
   text-align: center;
   color: $marigold;
